+++ app/src/main/AndroidManifest.xml
@@ -14,14 +14,14 @@
         android:supportsRtl="true"
         android:theme="@style/Theme.Tester"
         tools:targetApi="31">
-        <activity android:name=".RegisterUser" />
-        <activity android:name=".SigninUser" />
-        <activity android:name=".HomePageUI" />
-        <activity android:name=".PomodoroTimer"
+        <activity android:name=".UI.RegisterUser" />
+        <activity android:name=".UI.SigninUser" />
+        <activity android:name=".UI.HomePageUI" />
+        <activity android:name=".UI.PomodoroTimer"
             android:exported="true"/>
-        <activity android:name=".TaskListFragment"
+        <activity android:name=".UI.TaskListFragment"
             android:exported="true"/>
-        <activity android:name=".MainActivity"
+        <activity android:name=".UI.MainActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
