+++ app/src/main/java/com/example/tester/UI/CustomPrompt.java
@@ -1,4 +1,4 @@
-package com.example.tester;
+package com.example.tester.UI;
 
 import android.content.Context;
 import android.widget.Toast;
