+++ app/src/main/java/com/example/tester/UI/TimerFragment.java
@@ -1,4 +1,4 @@
-package com.example.tester;
+package com.example.tester.UI;
 
 import android.os.Bundle;
 
@@ -7,6 +7,8 @@ import androidx.fragment.app.Fragment;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
+
+import com.example.tester.R;
 
 /**
  * A simple {@link Fragment} subclass.
