+++ app/src/main/java/com/example/tester/entities/DatabaseHandler.java
@@ -1,12 +1,10 @@
-package com.example.tester.Utils;
+package com.example.tester.entities;
 
 import android.content.ContentValues;
 import android.content.Context;
 import android.database.Cursor;
 import android.database.sqlite.SQLiteDatabase;
 import android.database.sqlite.SQLiteOpenHelper;
-
-import com.example.tester.Model.TaskModel;
 
 import java.util.ArrayList;
 import java.util.List;
