+++ app/src/main/java/com/example/tester/exceptions/ValidationException.java
@@ -1,4 +1,4 @@
-package com.example.tester;
+package com.example.tester.exceptions;
 
 public class ValidationException extends Exception {
     public ValidationException(String message) {
