+++ app/src/main/res/layout/user_register.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".MainActivity"
+    tools:context=".UI.MainActivity"
     android:background="#F3ECE7"
     >
