+++ app/src/test/java/com/example/tester/TaskListFragmentTest.java
@@ -16,9 +16,10 @@ import androidx.fragment.app.FragmentManager;
 import androidx.fragment.app.FragmentTransaction;
 import androidx.recyclerview.widget.RecyclerView;
 
-import com.example.tester.Adapters.TaskAdapter;
-import com.example.tester.Model.TaskModel;
-import com.example.tester.Utils.DatabaseHandler;
+import com.example.tester.UI.TaskAdapter;
+import com.example.tester.entities.TaskModel;
+import com.example.tester.UI.TaskListFragment;
+import com.example.tester.entities.DatabaseHandler;
 import com.google.android.material.bottomnavigation.BottomNavigationView;
 import com.google.android.material.floatingactionbutton.FloatingActionButton;
